Python/requirements.txt
pyopencl
python-firebase
